+++ src/main/java/dev/ftb/mods/ftbic/events/CapabilityRegistrar.java
@@ -2,7 +2,7 @@ package dev.ftb.mods.ftbic.events;
 
 import dev.ftb.mods.ftbic.FTBIC;
 import dev.ftb.mods.ftbic.block.entity.machine.BatchFeederBlockEntity;
-import dev.ftb.mods.ftbic.block.entity.machine.CentrifugeBlockEntity;
+import dev.ftb.mods.ftbic.block.entity.machine.FluidMachineBlockEntity;
 import dev.ftb.mods.ftbic.FTBICConfig;
 import dev.ftb.mods.ftbic.block.ElectricBlockInstance;
 import dev.ftb.mods.ftbic.block.FTBICBlocks;
@@ -66,10 +66,10 @@ public final class CapabilityRegistrar {
 								((BatchFeederBlockEntity) be).fluidHandler));
 			}
 
-			if (instance == FTBICElectricBlocks.CENTRIFUGE || instance == FTBICElectricBlocks.ADVANCED_CENTRIFUGE) {
+			if (instance == FTBICElectricBlocks.CENTRIFUGE || instance == FTBICElectricBlocks.ADVANCED_CENTRIFUGE || instance == FTBICElectricBlocks.ORE_WASHER) {
 				event.registerBlockEntity(Capabilities.Fluid.BLOCK, type,
 						(be, side) -> new SidedResourceHandler<>(be, side, SideConfiguration.Resource.FLUIDS,
-								((CentrifugeBlockEntity) be).fluidHandler));
+								((FluidMachineBlockEntity) be).fluidHandler));
 			}
 
 			if (fullFE && instance.feCapMode != ElectricBlockInstance.FECapMode.INSERT_ONLY) {